backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
python-multipart==0.0.6
openai-whisper==20231117
transformers==4.35.2
TTS==0.22.0
spotipy==2.23.0
google-generativeai==0.3.1
pygetwindow==0.0.9
python-dotenv==1.0.0
sentencepiece
pydantic
soundfile
httpx
thefuzz==0.22.1
python-Levenshtein==0.25.0
pycaw
PyAutoGUI==0.9.54
pvporcupine==3.0.2
pyaudio==0.2.14 
requests==2.31.0
sounddevice==0.4.6
numpy==1.26.4
yt-dlp